requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio-dgram==2.1.2
attrs==21.4.0
chardet==4.0.0
click==8.1.3
discord.py==1.7.3
dnspython==2.2.1
idna==3.3
mcstatus==9.0.4
multidict==6.0.2
python-dotenv==0.20.0
typing_extensions==4.2.0
yarl==1.7.2